Split bot helpers into separate modules by concern
Move formatting utilities to bot_format.py and logging helpers to
bot_logging.py, leaving only core loop logic in bot.py. Circular
imports resolved via lazy imports and TYPE_CHECKING guards.
